increase journalctl collection to 7 days
I don't think 24 hours is enough time - sometimes we get a case, and it might be 24 hours before we get a response from "please run this host collector" and the important time window may have passed.  I might even prefer not using `--since` option at all and defaulting to collecting everything present in journald at the time.  But 7 might be a good compromise